+++ packages/admin/src/Http/Livewire/Components/Collections/CollectionTree.php
@@ -2,8 +2,8 @@
 
 namespace Lunar\Hub\Http\Livewire\Components\Collections;
 
-use Lunar\Facades\DB;
 use Livewire\Component;
+use Lunar\Facades\DB;
 use Lunar\Hub\Http\Livewire\Traits\MapsCollectionTree;
 use Lunar\Hub\Http\Livewire\Traits\Notifies;
 use Lunar\Models\Collection;