+++ packages/core/config/pricing.php
@@ -10,4 +10,18 @@ return [
     |
     */
     'stored_inclusive_of_tax' => false,
+  
+    /*
+    |--------------------------------------------------------------------------
+    | Pricing Pipelines
+    |--------------------------------------------------------------------------
+    |
+    | Define which pipelines should be run when retrieving purchasable price.
+    |
+    | Each pipeline class will be run from top to bottom.
+    |
+    */
+    'pipelines' => [
+        //
+    ],
 ];