+++ packages/admin/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-### [Unreleased]
+## [Unreleased]
+
+### Fixed
+
+- `values` are now detached from the product variant when it's the only one left.
 
 ### Changed